+++ design-system/storybook/stories/archive/atoms.story.jsx
@@ -7,7 +7,7 @@ import {
   BlockQuoteWrapper,
   KsButton,
   Details,
-  Select,
+  KsSelect,
   StatusMessage,
   Tooltip,
 } from '@knapsack/atoms';
@@ -19,19 +19,19 @@ storiesOf('Atoms', module)
     withInfo({
       inline: true,
     })(() => (
-      <Select
+      <KsSelect
         handleChange={action('option changed')}
         items={[
           {
-            title: 'Option 1',
+            label: 'Option 1',
             value: 'option1',
           },
           {
-            title: 'Option 2',
+            label: 'Option 2',
             value: 'option2',
           },
           {
-            title: 'Option 3',
+            label: 'Option 3',
             value: 'option3',
           },
         ]}