+++ CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.30](https://github.com/basaltinc/knapsack/compare/v2.0.0-beta.29...v2.0.0-beta.30) (2020-01-07)
+
+
+### Bug Fixes
+
+* provide react types in single export ([8ba66f7](https://github.com/basaltinc/knapsack/commit/8ba66f768500b01dc83b57ba01f2d488c94da15e))
+
+
+### Features
+
+* generate VS Code web component data files from spec ([9a2fa3b](https://github.com/basaltinc/knapsack/commit/9a2fa3b8693920de0ac364c396b5eea5c290bcbc))
+
+
+
+
+
+# [2.0.0-beta.29](https://github.com/basaltinc/knapsack/compare/v2.0.0-beta.28...v2.0.0-beta.29) (2020-01-07)
+
+**Note:** Version bump only for package @knapsack/monorepo
+
+
+
+
+
+# [2.0.0-beta.28](https://github.com/basaltinc/knapsack/compare/v2.0.0-beta.27...v2.0.0-beta.28) (2020-01-06)
+
+
+### Bug Fixes
+
+* new patterns get first status ([bd21924](https://github.com/basaltinc/knapsack/commit/bd21924a645c6e974ad07c4184afa09e972e6473))
+
+
+
+
+
+# [2.0.0-beta.27](https://github.com/basaltinc/knapsack/compare/v2.0.0-beta.26...v2.0.0-beta.27) (2020-01-06)
+
+
+### Features
+
+* rich selects ([#351](https://github.com/basaltinc/knapsack/issues/351)) ([427bcb6](https://github.com/basaltinc/knapsack/commit/427bcb6b2cfb08856d63ac3012345445030a06fa))
+
+
+
+
+
 # [2.0.0-beta.26](https://github.com/basaltinc/knapsack/compare/v2.0.0-beta.25...v2.0.0-beta.26) (2020-01-03)
 
 **Note:** Version bump only for package @knapsack/monorepo